shuffle the storage
- providers in their own files
- send a storage instance into charm manager
- configuration types
