+++ typescript/packages/jumble/src/contexts/CharmManagerContext.tsx
@@ -1,6 +1,7 @@
 import React, { createContext, useContext, useEffect, useRef, useState } from "react";
 import { CharmManager } from "@commontools/charm";
 import { useParams } from "react-router-dom";
+import { createStorage } from "@commontools/charm";
 
 export type CharmManagerContextType = {
   charmManager: CharmManager;
@@ -9,7 +10,7 @@ export type CharmManagerContextType = {
 
 const CharmManagerContext = createContext<CharmManagerContextType>(null!);
 
-const defaultManager = new CharmManager(undefined, "memory");
+const defaultManager = new CharmManager(createStorage({ type: "memory" }));
 
 export const CharmsManagerProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const { replicaName } = useParams<{ replicaName: string }>();
@@ -26,7 +27,10 @@ export const CharmsManagerProvider: React.FC<{ children: React.ReactNode }> = ({
 
     // Create new charm manager instance with updated replica
     const storageType = (import.meta as any).env.VITE_STORAGE_TYPE ?? "remote";
-    const manager = new CharmManager(effectiveReplica, storageType);
+    const storage = storageType === "remote" ?
+      createStorage({ type: "remote", replica: effectiveReplica, url: new URL(location.href) }) :
+      createStorage({ type: storageType as "memory" | "local" });
+    const manager = new CharmManager(storage);
     setCharmManager(manager);
   }, [effectiveReplica]);
